+++ lib/generators/config/templates/config.rb
@@ -15,6 +15,10 @@ Config.setup do |config|
   #
   # config.overwrite_arrays = true
 
+  # Defines current environment, affecting which settings file will be loaded.
+  # Default: `::Rails.env`
+  # config.environment = ENV.fetch('ENVIRONMENT', :development)
+
   # Load environment variables from the `ENV` object and override any settings defined in files.
   #
   # config.use_env = false