+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 
 ### Changes
 
+* Add option to specify environment (default: `RAILS_ENV`) ([#290](https://github.com/rubyconfig/config/pull/290))
 * Add JRuby 9.2 to the test matrix ([#228](https://github.com/railsconfig/config/issues/228))
 * Add exit! to reserved keywords ([#289](https://github.com/railsconfig/config/issues/289))